+++ src/components/Button/styles.ts
@@ -1,0 +1,12 @@
+import styled from 'styled-components';
+
+export const Button = styled.button`
+  line-height: 0;
+  border-radius: 50%;
+  cursor: pointer;
+`;
+
+export const Icon = styled.img`
+  width: 2rem;
+  height: 2rem;
+`;
